+++ amis/build_env_setup.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# This script sets up environment for building cfncluster ami
+# This script sets up environment for building AWS ParallelCluster ami
 # Launch EC2 machine with amazon linux and run this script
 
 # Get latest packages and Development tools
@@ -20,10 +20,10 @@ unzip packer_1.2.4_linux_amd64.zip
 sudo cp packer /usr/local/bin/
 packer --version
 
-# Get cfncluster cookbook - update to your own fork in case
+# Get AWS ParallelCluster cookbook - update to your own fork in case
 # if you are testing your changes
-# git clone https://github.com/awslabs/cfncluster-cookbook.git
-# cd cfncluster-cookbook
+# git clone https://github.com/awslabs/aws-parallelcluster-cookbook.git
+# cd aws-parallelcluster-cookbook
 
 # Run ami builder script
 # ./build_ami.sh <operating_system> <region> <public/private>