+++ recipes/_nvidia_install.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: cfncluster
+# Cookbook Name:: aws-parallelcluster
 # Recipe:: _nvidia_install
 #
 # Copyright 2013-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.